--- bci-game/Assets/playerHealth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b38417abb9134f741a5b1e63979c2030
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
